# mat: format complex dense matrices