--- a/mat/format_value.go
+++ b/mat/format_value.go
@@ -1,0 +1,143 @@
+// Copyright ©2021 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package mat
+
+import (
+	"fmt"
+	"strconv"
+	"sync"
+)
+
+const (
+	syntaxDefault = iota
+	syntaxMATLAB
+	syntaxPython
+
+	defaultDotByte = '.'
+)
+
+// valueFormatter is a type that can set options for formatting values and that
+// satisfies the fmt.Formatter interface.
+type valueFormatter interface {
+	fmt.Formatter
+
+	setDotByte(b byte)
+	setSyntax(s int)
+}
+
+// vFormatOption is a functional option for value formatting.
+type vFormatOption func(valueFormatter)
+
+// vDotByte sets the dot character to b. The dot character is used to replace a
+// value if the result is printed with the fmt ' ' verb flag. Without a DotByte
+// option, the default dot character is '.'.
+func vDotByte(b byte) vFormatOption {
+	return func(f valueFormatter) { f.setDotByte(b) }
+}
+
+// vFormatMATLAB sets the printing behavior to output MATLAB syntax. If MATLAB
+// syntax is specified, the ' ' verb flag is ignored.
+func vFormatMATLAB() vFormatOption {
+	return func(f valueFormatter) { f.setSyntax(syntaxMATLAB) }
+}
+
+// vFormatPython sets the printing behavior to output Python syntax. If Python
+// syntax is specified, the ' ' verb flag is ignored.
+func vFormatPython() vFormatOption {
+	return func(f valueFormatter) { f.setSyntax(syntaxPython) }
+}
+
+// formattedFloat returns a fmt.Formatter for the floating point value v using
+// the given options.
+func formattedFloat(v float64, options ...vFormatOption) fmt.Formatter {
+	f := floatFormatter{
+		value:  v,
+		buf:    make([]byte, 0, 64),
+		dot:    defaultDotByte,
+		syntax: syntaxDefault,
+	}
+	for _, o := range options {
+		o(&f)
+	}
+	return &f
+}
+
+// floatFormatter formats 64-bit a floating point value and satifies the
+// valueFormatter interface. A floatFormatter utilizes an internal buffer while
+// generating formatting output and this buffer persists with the formatter.
+// For this reason, and because a floatFormatter might be reused or revalued,
+// the floatFormatter embeds an exclusion lock.
+type floatFormatter struct {
+	sync.Mutex
+
+	value  float64
+	buf    []byte
+	dot    byte
+	syntax int
+
+	format func(v float64, buf []byte, dot byte, syntax int, fs fmt.State, c rune)
+}
+
+var _ valueFormatter = (*floatFormatter)(nil)
+
+// Format satisfies the fmt.Formatter interface.
+func (f *floatFormatter) Format(fs fmt.State, c rune) {
+	f.Lock()
+	defer f.Unlock()
+
+	if f.format == nil {
+		f.format = formatFloat
+	}
+	f.format(f.value, f.buf, f.dot, f.syntax, fs, c)
+}
+
+func (f *floatFormatter) setDotByte(b byte) { f.dot = b }
+func (f *floatFormatter) setSyntax(s int)   { f.syntax = s }
+
+// formatFloat prints a representation of v to the fs io.Writer.  The format
+// character c specifies the numerical representation; valid values are those
+// for float64 specified in the fmt package, with their associated flags. In
+// addition to this, a space preceding a verb indicates that zero values should
+// be represented by the dot character.
+//
+// formatFloat will not provide Go syntax output
+func formatFloat(v float64, buf []byte, dot byte, syntax int, fs fmt.State, c rune) {
+	buf = buf[:0]
+
+	if (v == 0) && (syntax == syntaxDefault) && fs.Flag(' ') {
+		buf = append(buf, dot)
+	} else {
+		prec, ok := fs.Precision()
+		if !ok {
+			prec = -1
+		}
+
+		if c == 'v' {
+			c = 'g'
+		}
+
+		if (v >= 0) && fs.Flag('+') {
+			buf = append(buf, '+')
+		}
+
+		buf = strconv.AppendFloat(buf, v, byte(c), prec, 64)
+	}
+
+	width, ok := fs.Width()
+	if ok && (width > len(buf)) {
+		if fs.Flag('-') {
+			fs.Write(buf)
+			for i := 0; i < width-len(buf); i++ {
+				fmt.Fprintf(fs, " ")
+			}
+			return
+		}
+		for i := 0; i < width-len(buf); i++ {
+			fmt.Fprintf(fs, " ")
+		}
+	}
+
+	fs.Write(buf)
+}