Please take a look.

This PR adds a `CFormat` function in gonum/mat to format complex dense matrices; this function is intended to complement the current `Format` function.  Relevant issue #641 and a relevant PR #1414, although that PR appears to have stalled.  Further, PR #1414 takes the approach of wholesale duplicating `format.go`, which may not be preferred.

Summary:

Following https://github.com/gonum/gonum/pull/1414#issuecomment-666282217, an objective was to avoid code duplication while implementing a formatter for complex matrices.  To summarize the proposed approach:  value formatters are added for formatting floats and complex numbers, which are then used during matrix formatting.  Since these value formatters satisfy the `fmt.Formatter` interface, `fmt.State` can be passed, offering simplification in several areas (e.g., re-construction of a format string is no longer necessary).  As it would be inefficient to spin up a a new formatter for each element in a matrix, a value formatter holds a buffer for reuse.

To tackle widthing, a `state` type is proposed (satisfying `fmt.State`) that can be fit to a matrix and that stores information on column widths, using the existing widther type. As before, iterating through matrix elements is required.   After a call to `At`, a `state` will respond to subsequent `Width` calls with the relevant column width.  Recommendations toward improving implementation are appreciated.

Notwithstanding the need for additional internal types, this feels simpler to me on the whole and I believe this approach should make extending to new syntaxes or data types easier, if desired.  Thanks for reviewing -- will address feedback/comments. 

Bench:

I've included a few benchmarks below.  These compare the current and proposed implementation when formatting 10x10, 100x100 and 1000x1000 dense matrices.

| implementation | benchmark           |  iter |            time |          bytes |           alloc |
| :------------- | :------------------ | ----: | --------------: | -------------: | --------------: |
| current        | BenchmarkFormat10   | 36134 |     33063 ns/op |      2693 B/op |    16 allocs/op |
| current        | BenchmarkFormat100  |   358 |   3747403 ns/op |   1672025 B/op |   137 allocs/op |
| current        | BenchmarkFormat1000 |     3 | 334034817 ns/op | 163557800 B/op |  1066 allocs/op |
| proposed       | BenchmarkFormat10   | 22086 |     54554 ns/op |      2886 B/op |    19 allocs/op |
| proposed       | BenchmarkFormat100  |   170 |   7072273 ns/op |   1670752 B/op |    49 allocs/op |
| proposed       | BenchmarkFormat1000 |     2 | 697675398 ns/op | 163542896 B/op |    80 allocs/op |