@wamuir 2021-08-13
> just a drive-by comment:
> could you run the benchmarks before/after the proposed optimization a couple of times (say 10-20) and show the result of `benchstat`?
> with the limited statistics you presented in the commit message, I get this rather poor statistical power result:
> 
> ```
> $> benchstat ./ref.txt ./new.txt 
> name        old time/op    new time/op    delta
> Format10      33.1µs ± 0%    54.6µs ± 0%   ~     (p=1.000 n=1+1)
> Format100     3.75ms ± 0%    7.07ms ± 0%   ~     (p=1.000 n=1+1)
> Format1000     334ms ± 0%     698ms ± 0%   ~     (p=1.000 n=1+1)
> 
> name        old alloc/op   new alloc/op   delta
> Format10      2.69kB ± 0%    2.89kB ± 0%   ~     (p=1.000 n=1+1)
> Format100     1.67MB ± 0%    1.67MB ± 0%   ~     (p=1.000 n=1+1)
> Format1000     164MB ± 0%     164MB ± 0%   ~     (p=1.000 n=1+1)
> 
> name        old allocs/op  new allocs/op  delta
> Format10        16.0 ± 0%      19.0 ± 0%   ~     (p=1.000 n=1+1)
> Format100        137 ± 0%        49 ± 0%   ~     (p=1.000 n=1+1)
> Format1000     1.07k ± 0%     0.08k ± 0%   ~     (p=1.000 n=1+1)
> ```
> 
> ie:
> 
> ```
> $> go test -run=NONE -bench=Format -benchmem -count=20 |& tee ref.txt
> ## apply optimizations
> $> go test -run=NONE -bench=Format -benchmem -count=20 |& tee new.txt
> $> benchstat ./ref.txt ./new.txt
> ```
> 
> thanks.
> 
> PS: `benchstat` is available here:
> 
>     * [golang.org/x/perf/benchstat](https://golang.org/x/perf/benchstat)

Thanks, I wasn't aware of benchstat.  Cool.  Output is below.  I added `-benchtime=1x`, because this appears to otherwise perform tests using the pooled mean?  If true, I'd think the dispersion of ± time/op is otherwise understated. I guess that's a nit on `go test` for not having a flag to output variance (or a nit on me for not being able to find it).

| name       | old time/op | new time/op  | delta                       |
| :--------- | ----------: | -----------: | --------------------------: |
| Format10   | 44.2µs ±17% |  73.9µs ±38% |  +67.07%  (p=0.000 n=20+18) |
| Format100  | 3.28ms ± 3% |  7.55ms ± 5% | +130.21%  (p=0.000 n=20+18) |
| Format1000 |  329ms ± 1% |   764ms ± 2% | +132.17%  (p=0.000 n=20+20) |

Increase in clock time is statistically significant.  @kortschak provided review comments, which I'll work through over the weekend and then rerun the bench.  One of these comments is to use micro-benchmarking and I expect this will offer greater insight into differences.  Addressing his comments may also close the current gap (e.g., his comment regarding fmt.Fprint and reflect).