@wamuir 2021-08-14
Most recent commit addresses review comments, except for the switch to sub-benchmarks.  I'll need to push a second commit to update the benchmarks.  I attempted some further simplification of the format functions and I think this was productive.  I also added comments along the way.  Updated benchstat output is below.

| name       | old time/op | new time/op  | delta                       |
| :--------- | ----------: | -----------: | --------------------------: |
| Format10   | 44.9µs ±10% |  62.9µs ±31% |  +40.08%  (p=0.000 n=18+20) |
| Format100  | 3.31ms ± 1% |  4.27ms ± 1% |  +29.20%  (p=0.000 n=18+20) |
| Format1000 |  326ms ± 1% |   426ms ± 1% |  +30.41%  (p=0.000 n=20+20) |