@wamuir 2021-08-14
| name                | old time/op | new time/op | delta                      |
| :------------------ | ----------: | ----------: | -------------------------: |
| Format/10/General   | 52.6µs ±17% | 68.8µs ±18% | +30.74%  (p=0.000 n=19+19) |
| Format/10/DotByte   | 49.5µs ±21% | 67.9µs ±17% | +37.19%  (p=0.000 n=19+20) |
| Format/10/Excerpt   | 50.3µs ±11% | 66.1µs ±21% | +31.40%  (p=0.000 n=19+20) |
| Format/10/Prefix    | 49.4µs ±10% | 66.4µs ±17% | +34.54%  (p=0.000 n=17+20) |
| Format/10/Squeeze   | 53.0µs ±22% | 69.7µs ±14% | +31.49%  (p=0.000 n=20+20) |
| Format/10/MATLAB    | 51.5µs ±12% | 66.4µs ±12% | +28.79%  (p=0.000 n=20+19) |
| Format/10/MATLAB#   | 43.0µs ±18% | 42.9µs ±23% |    ~     (p=0.967 n=20+20) |
| Format/10/Python    | 50.4µs ±13% | 64.0µs ± 9% | +26.99%  (p=0.000 n=20+19) |
| Format/10/Python#   | 42.4µs ±37% | 43.1µs ±30% |    ~     (p=0.757 n=20+20) |
| Format/100/General  | 3.13ms ± 7% | 4.43ms ± 9% | +41.42%  (p=0.000 n=17+20) |
| Format/100/DotByte  | 3.94ms ± 1% | 4.35ms ± 8% | +10.32%  (p=0.000 n=19+20) |
| Format/100/Excerpt  | 4.00ms ± 1% | 4.54ms ± 6% | +13.45%  (p=0.000 n=20+20) |
| Format/100/Prefix   | 4.00ms ± 1% | 5.61ms ± 1% | +40.23%  (p=0.000 n=20+16) |
| Format/100/Squeeze  | 4.42ms ± 1% | 5.35ms ±18% | +21.01%  (p=0.000 n=14+20) |
| Format/100/MATLAB   | 4.68ms ±15% | 6.46ms ±13% | +37.92%  (p=0.000 n=20+20) |
| Format/100/MATLAB#  | 2.91ms ± 3% | 2.90ms ± 2% |    ~     (p=0.247 n=20+20) |
| Format/100/Python   | 6.33ms ±16% | 6.38ms ±35% |    ~     (p=0.890 n=20+20) |
| Format/100/Python#  | 3.56ms ±13% | 1.69ms ± 1% | -52.45%  (p=0.000 n=20+19) |
| Format/1000/General |  306ms ± 0% |  405ms ± 0% | +32.08%  (p=0.000 n=19+20) |
| Format/1000/DotByte |  303ms ± 0% |  392ms ± 0% | +29.54%  (p=0.000 n=20+20) |
| Format/1000/Excerpt |  308ms ± 1% |  405ms ± 0% | +31.76%  (p=0.000 n=20+20) |
| Format/1000/Prefix  |  307ms ± 0% |  404ms ± 0% | +31.67%  (p=0.000 n=20+20) |
| Format/1000/Squeeze |  307ms ± 0% |  405ms ± 0% | +31.83%  (p=0.000 n=20+20) |
| Format/1000/MATLAB  |  307ms ± 0% |  405ms ± 0% | +31.99%  (p=0.000 n=19+20) |
| Format/1000/MATLAB# |  164ms ± 1% |  164ms ± 1% |    ~     (p=0.561 n=19+19) |
| Format/1000/Python  |  308ms ± 1% |  406ms ± 1% | +31.86%  (p=0.000 n=20+18) |
| Format/1000/Python# |  164ms ± 1% |  164ms ± 1% |    ~     (p=0.604 n=20+19) |
| [Geo mean]          |      3.74ms |      4.43ms | +18.29%                    |